Rename EditLineForm -> LineForm
"Edit" prefix doesn't add any useful information about purpose
of the form
